# openapi.yaml
openapi: 3.1.0
info:
  title: SpecR API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
  description: >
    Headless REST API for CSI® MasterFormat®-compatible specification document
    automation with round-trip DOCX support. Independent project; not affiliated
    with CSI.

# No authentication yet — auth lands with Phase 5f (#43). The explicit empty
# security array documents that every endpoint is currently unauthenticated.
security: []
servers:
  - url: http://localhost:3000
    description: Local development server

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /specs/{id}/generate:
    post:
      operationId: generateDocx
      summary: Generate a DOCX from a stored spec AST
      tags: [specs]
      parameters:
        - $ref: '#/components/parameters/SpecId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                templateId:
                  type: string
                  format: uuid
                  description: >
                    Style template to apply (font, spacing, indent, numbering format).
                    Omitted → the seeded UFGS-Default template is applied; if that template
                    is absent (un-migrated database) the response is still 200 with an
                    unstyled DOCX. Unknown id → 404.
                sectionNumberFormat:
                  type: string
                  enum: [canonical, dots, compact]
                  default: canonical
                  description: >
                    Presentation format for generated section numbers in the synthetic
                    DOCX title and confidently identified Section references in paragraph
                    text. Canonical emits "09 91 00", dots emits "09.91.00", and compact
                    emits "099100". Stored AST and database values remain canonical.
            examples:
              dotted-section-numbers:
                value:
                  sectionNumberFormat: dots
      responses:
        '200':
          description: DOCX file stream
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /projects/{id}/generate:
    post:
      operationId: generateManual
      summary: Assemble a project manual DOCX from all TOC sections
      description: >
        Renders every section in the project's TOC (MasterFormat / position order)
        into a single DOCX manual (ADR-017 D1). Each section is its own OOXML
        section with an independent numbering instance, so multilevel numbering
        restarts at PART 1 per section. Every paragraph keeps its `w:sdt` UUID
        anchor for round-trip. The manual opens with a cover page (project name +
        description, style-template applied) and a Word TOC field over the section
        headings — Word computes the TOC entries and pagination on open; SpecR emits
        the field, not page numbers. This route renders the live draft from
        `project_specs`; issued and addendum manuals render from frozen package
        revisions via `POST /revisions/{id}/generate`.
      tags: [projects]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                templateId:
                  type: string
                  format: uuid
                  description: >
                    Style template applied to every section. Omitted → the seeded
                    UFGS-Default template; absent default → unstyled 200. Unknown id → 404.
                sectionNumberFormat:
                  type: string
                  enum: [canonical, dots, compact]
                  default: canonical
                  description: >
                    Presentation format for generated section numbers in each
                    section title and confidently identified Section references.
      responses:
        '200':
          description: DOCX manual file stream
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Project has no sections to assemble
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /templates/{id}:
    get:
      operationId: getTemplate
      summary: Retrieve a style template with its rules
      tags: [templates]
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '200':
          description: Template with rules
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/Template'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      operationId: patchTemplate
      summary: Update template metadata (name and/or owner)
      description: >
        At least one of `name` or `owner` must be present. Pass `owner: null`
        to clear the owner field.
      tags: [templates]
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  type: string
                  minLength: 1
                owner:
                  type: [string, 'null']
                  minLength: 1
                  description: Pass null to clear the owner
      responses:
        '200':
          description: Updated template metadata
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        $ref: '#/components/schemas/TemplateMeta'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteTemplate
      summary: Delete a style template and all its rules
      description: >
        A template assigned to any spec as its style source cannot be deleted
        (ON DELETE RESTRICT, #138) — the request returns 409 with the count of
        referencing specs. Clear the association on those specs first
        (DELETE /specs/{id}/style-source).
      tags: [templates]
      parameters:
        - $ref: '#/components/parameters/TemplateId'
      responses:
        '204':
          description: Template deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Template is in use by one or more specs as a style source and cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    SpecId:
      name: id
      in: path
      required: true
      description: Spec UUID
      schema:
        type: string
        format: uuid
...
    ProjectId:
      name: id
      in: path
      required: true
      description: Project UUID
      schema:
        type: string
        format: uuid
...
    TemplateId:
      name: id
      in: path
      required: true
      description: Template UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request — validation error or unsupported file type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict — a resource with that identity already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string

    # ── Parse ──────────────────────────────────────────────────────────────────
...
    StyleNodeType:
      type: string
      enum: [part, article, pr1, pr2, pr3, pr4, pr5, pr6, pr7]
      description: >
        CSI paragraph role. Maps to the nine generator levels: `part` (PART 1 –
        GENERAL), `article` (1.1 REFERENCES), `pr1`–`pr7` (A. / 1. / a. / 1) /
        a), then repeated 1) / a) paragraph tiers).
    StyleRule:
      type: object
      required: [nodeType, properties]
      properties:
        nodeType:
          $ref: '#/components/schemas/StyleNodeType'
        properties:
          $ref: '#/components/schemas/StyleProperties'
    StyleProperties:
      type: object
      description: >
        OOXML-derived style properties for a CSI paragraph role. Recognized keys
        are documented below; **unknown OOXML keys are preserved** via the JSONB
        catchall schema. All sub-objects also accept additional OOXML properties.
      properties:
        rPr:
          type: object
          description: Run (character) properties
          properties:
            rFonts:
              type: object
              description: Font references
              properties:
                ascii:
                  type: string
                hAnsi:
                  type: string
                cs:
                  type: string
                eastAsia:
                  type: string
              additionalProperties: true
            sz:
              type: integer
              description: Font size in half-points (e.g. 20 = 10 pt)
            b:
              type: boolean
              description: Bold
            i:
              type: boolean
              description: Italic
            caps:
              type: boolean
              description: All-caps
            smallCaps:
              type: boolean
            u:
              type: string
              description: Underline type (e.g. "single")
            strike:
              type: boolean
            color:
              type: string
              description: 'RRGGBB hex or "auto" (e.g. "FF0000")'
            highlight:
              type: string
          additionalProperties: true
        pPr:
          type: object
          description: Paragraph properties
          properties:
            spacing:
              type: object
              properties:
                before:
                  type: integer
                  description: Space before in twips
                after:
                  type: integer
                  description: Space after in twips
                line:
                  type: integer
                  description: Line spacing value in twips
                lineRule:
                  type: string
                  enum: [auto, exact, atLeast]
                contextualSpacing:
                  type: boolean
              additionalProperties: true
            ind:
              type: object
              description: Indentation (twips)
              properties:
                left:
                  type: integer
                right:
                  type: integer
                firstLine:
                  type: integer
                hanging:
                  type: integer
              additionalProperties: true
            jc:
              type: string
              enum: [left, center, right, both, distribute, start, end]
          additionalProperties: true
        numbering:
          type: object
          description: Numbering definition (from `numbering.xml` level)
          properties:
            ilvl:
              type: integer
            numFmt:
              type: string
            lvlText:
              type: string
            start:
              type: integer
          additionalProperties: true
      additionalProperties: true
    TemplateMeta:
      type: object
      required: [id, name, owner, createdAt]
      description: Template identity and metadata without style rules
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Unique template name
        owner:
          type: [string, 'null']
          description: Firm / owner label, or null if not provided
        createdAt:
          type: string
          format: date-time
    Template:
      allOf:
        - $ref: '#/components/schemas/TemplateMeta'
        - type: object
          required: [rules]
          properties:
            rules:
              type: array
              items:
                $ref: '#/components/schemas/StyleRule'
              description: One rule per styleable NodeType (9 max)
...